fix: fix the unique name generation seed managment
The generator does not regenerate the seed at each call because it is created once at the obsidian launch. This means the seed is the date when Obsidian has been opened, that is not the expected behavior.

The generator implementation has been shifted into a UniqueNameGeneratorSeed containing the whole incrementale name generation logique. This way of shifting the concern makes the generator stateless as suggested in the recommendations.

BEFORE:

interface IUniqueNameGenerator {
    generateFromNow(): string;
}

AFTER:

interface IUniqueNameGenerator {
    generateNewSeed(): IUniqueNameGeneratorSeed;
}

export interface IUniqueNameGeneratorSeed {
    next: () => string;
}