chore(build): webpack bundling now targets node environment
Until now, I was using the polyfill mecanism to be able to use the node modules in my scripts and
export them.

This is the configuration I had in the webpack.base.config.js for the "path" module:

  module.exports = {
    ...
    resolve: {
      ...
      fallback: {
        path: require.resolve('path-browserify')
      }
    }
    ...
  }

In the new koboHighlightsImporter, I need to use the "fs" module and struggled to find a polyfill.
Trying to solve my problem with the node-polyfill-webpack-plugin package, I read this :

  The fs module resolves to nothing because its functionality cannot replicated in the browser.

What I forgot was that polyfill have been made to reproduce modern behavior in a browser
environment.
But in my case, Obsidian is in a node environment.

The answer is actually here : https://webpack.js.org/concepts/targets/

  module.exports = {
    target: 'node',
  };

Targeting node makes the node modules available in the bundled code.