feat(kobohighlightsimporter): implement the KoboHighlightsImporter QuickAdd script (#39)
This script mainly focus on importing into Obsidian the annotations taken into a Kobo.
A QuickAdd script can be seen has a box having inputs and outputs stored into the params.variables.
At this stage, it provides three outputs :
- The book title
- The book author
- The book annotations already in formated content
The script handles the books suggestion popup as entry, but this part is planned to be extracted
into a specific script.
This is the aim of the ticket #45.

Implemeting this script has been chalenging and this feature comes with a lot of changes in the
core part of the framework.

BREAKING CHANGE:

Project configuration
- deactivate @typescript-eslint/no-explicit-any lint is a bad idea. using any will not raise type
error. It has been a problem when I dealed with the formatter, I was not able to see the object I
gave was badly formed.
- add snippets for it.todo.

Core and Infrastructure implementations
- implement a FileSystem service used to get the files contained into a folder.
- implement an XmlParser service.
- UniqueNoteCreator is extracted from the UserPlugins NewUniqueNoteCommand.
- implement services and utility classes to handle Adobe Digital Editions annotations files.
- modify the SettingableScript architecture. The interface becomes the entry point and the
implementation is supported by an inner script.