chore(build): use windows images in the pipeline to fix the endline issue (#50)
Default endlines are not the same in Linux and Windows.
The pipeline is using ubuntu images forcing me to make the files end with "lf"
in the .editorconfig, otherwise the pipeline crash because of prettier that is
using the endline defines in the .editorconfig.

The fact is, I'm under Windows and the UI Git tool I use (Fork) is putting a
Windows endline to the source files each time I do an operation with. It's
really anoying. I don't want that anymore.