feat(quickadd): implement builders for the settings (#48)
Here are the main downsides of the default QuickAdd settings system :
- the UI option label is the name of the option root property,
- the settings given to the script at runtime is not strongly typed.

The way settings are working can be seen at the following URL :
https://quickadd.obsidian.guide/docs/Advanced/scriptsWithSettings

This ticket cover the implementation of two builders, each one covering one downside :

The SettingsDefinitionBuilder covers the first downside.
It introduces the 'name' property allowing the developer to reference a field by name
and no more by label.

  const SettingsDefinition = new SettingsDefinitionBuilder()
    .forScript('Hello World')
    .implementedBy('me')
    .havingTextFieldOption({
        name: nameof<HelloWorldSettings>((s) => s.name),
        label: 'Name to say hello',
        defaultValue: 'World'
    })
    .build();

The SettingsBuilder covers the second downside.
The developer can now define its settings using an interface.
Given the native settings and the previously defined definition, this builder is able
to build the strongly typed settings.

  interface HelloWorldSettings {
    name: string;
  }

  const settingsBuilder = new SettingsBuilder<HelloWorldSettings>();
  const stronglyTypedSettings = settingsBuilder.build(settings, definition);

Documentation has been updated accordingly.

BREAKING CHANGE:

QuickAdd scripts are not using the native settings system anymore. It is now based on specific
SettingsDefinitionBuilder and SettingsBuilder builders.