chore: using husky and commintlint only on [*/]main branches (#49)
The configured hooks always ran. I've been mainly avoiding them during my work in progress
because it tooks too much time to execute.

The fact is, I need to be sure the commit message is well formated and the pipeline
succeed mainly at push time on a next/main or main branches. Means, before the code goes
to the server.

I as well kept the commit-msg and post-rewrite hooks for the main branches to have an
immediate warning in case I need to directly work on them.