feat(quickadd): implement a files suggester (#45)
The files suggester code block embedded into the Kobo Highlights Extrator has been
extracted to become a script on its own. This way, it can be used by other macros.

Its behaviour is simple:
- it takes a directory in parameter (settings),
- and suggest to the end user the level 0 list of files this directory contains.

The file selected by the end user is set into a QUickAdd variable in order to
be used by an other step.