feat(quickadd): variables mapper becomes variable formatter (#56)
The format setting is added to the script making it a formatter and not a mapper anymore.
This new naming adds value to the script.

This settings property allows the user to format the value using the handlebars plugin.
The object given to the formatter fits to the following interface.

  {
    value: unknow;
  }

If you want to format a string and transform it to a link, here is then the value of the format
property:

  [[{{value}}]]

Whatever is the input value type, the output is a string.