docs(quickadd): implement the new task note macro
When I work on my projects, sometimes a simple task is enough to follow it.
But when the task becomes bigger and specially when I need to take note relatively, I want to be able to
quickly transform it to a note and still have the link to the parent without losing the description.